Only use "allow listed" app names in the Fission namespace
