+++ fission-cli/library/Fission/CLI/Parser/Command/App/Init/Types.hs
@@ -2,18 +2,18 @@ module Fission.CLI.Parser.Command.App.Init.Types (Options (..)) where
 
 import           Fission.Prelude
 
-import qualified Fission.URL                          as URL
+import qualified Fission.App.Name                     as App
 
 import qualified Fission.CLI.Parser.Config.IPFS.Types as IPFS
 import           Fission.CLI.Parser.Verbose.Types
 
 -- | Arguments, flags & switches for the `app init` command
 data Options = Options
-  { appDir       :: FilePath
-  , buildDir     :: Maybe FilePath
-  , maySubdomain :: Maybe URL.Subdomain
-  , ipfsCfg      :: IPFS.Config
-  , verboseFlag  :: VerboseFlag -- ^ Verbose flag
+  { appDir      :: FilePath
+  , buildDir    :: Maybe FilePath
+  , mayAppName  :: Maybe App.Name
+  , ipfsCfg     :: IPFS.Config
+  , verboseFlag :: VerboseFlag -- ^ Verbose flag
   } deriving (Show, Eq)
 
 instance Has VerboseFlag Options where
