+++ fission-core/library/Fission/App/Name/Error.hs
@@ -1,0 +1,9 @@
+module Fission.App.Name.Error where
+
+import           Fission.Prelude
+
+data Invalid = Invalid
+  deriving (Eq, Show, Exception)
+
+instance Display Invalid where
+  display Invalid = "Invalid App.Name -- must be alphanumeric separated with hyphens and not blocklisted"
