+++ fission-core/library/Fission/URL/Subdomain/Types.hs
@@ -1,24 +1,19 @@
 module Fission.URL.Subdomain.Types (Subdomain (..)) where
 
-import qualified RIO.ByteString.Lazy  as Lazy
-import qualified RIO.Text             as Text
+import qualified RIO.ByteString.Lazy    as Lazy
+import qualified RIO.Text               as Text
 
-import           Data.Swagger         hiding (get)
-import           Database.Persist.Sql hiding (get)
-
+import           Data.Swagger
+import           Database.Persist.Sql
 import           Servant.API
 
 import           Fission.Prelude
 
--- | Type safety wrapper for subdomains
-newtype Subdomain = Subdomain { get :: Text }
-  deriving          ( Eq
-                    , Show
-                    )
-  deriving newtype  ( IsString )
+import qualified Fission.App.Name.Types as App
 
-instance Display Subdomain where
-  textDisplay (Subdomain txt) = txt
+-- | Type safety wrapper for subdomains
+newtype Subdomain = Subdomain { raw :: Text }
+  deriving newtype (Eq , Show, Display)
 
 instance Semigroup Subdomain where
   Subdomain subA <> Subdomain subB = Subdomain (subA <> "." <> subB)
@@ -58,10 +53,10 @@ instance ToHttpApiData Subdomain where
   toUrlPiece = textDisplay
 
 instance MimeRender PlainText Subdomain where
-  mimeRender _ = displayLazyBS . get
+  mimeRender _ = displayLazyBS . raw
 
 instance MimeRender OctetStream Subdomain where
-  mimeRender _ = displayLazyBS . get
+  mimeRender _ = displayLazyBS . raw
 
 instance MimeUnrender PlainText Subdomain where
   mimeUnrender _proxy bs =
@@ -72,226 +67,5 @@ instance MimeUnrender PlainText Subdomain where
 
 instance Arbitrary Subdomain where
   arbitrary = do
-    generators <- sublistOf [opinions, sizes, ages, shapes, colours, materials]
-
-    if null generators
-      then
-        arbitrary
-
-      else do
-        adjectives <- sequence (elements <$> take 3 generators)
-        noun       <- elements nouns
-        return . Subdomain . Text.intercalate "-" $ adjectives <> [noun]
-
-opinions :: [Text]
-opinions =
-  [ "amazing"
-  , "beautiful"
-  , "ugly"
-  , "quick"
-  , "wonderful"
-  , "awesome"
-  , "sweet"
-  , "tubular"
-  , "eager"
-  , "magnificient"
-  , "nice"
-  , "lively"
-  , "bewildered"
-  , "fierce"
-  , "jolly"
-  , "victorious"
-  , "calm"
-  , "brave"
-  , "proud"
-  , "fancy"
-  , "skinny"
-  , "bald"
-  , "elegant"
-  , "muscular"
-  , "rich"
-  , "tasty"
-  , "super"
-  , "brainy"
-  , "infantile"
-  , "juvenile"
-  , "spry"
-  , "playful"
-  , "loyal"
-  , "vicious"
-  , "cute"
-  , "benevolent"
-  , "malevolent"
-  , "universal"
-  , "narcissistic"
-  , "wise"
-  , "stupid"
-  , "obtuse"
-  , "fun"
-  , "charming"
-  , "wicked"
-  , "gnarly"
-  ]
-
-sizes :: [Text]
-sizes =
-  [ "big"
-  , "small"
-  , "little"
-  , "huge"
-  , "enormous"
-  , "petite"
-  , "tall"
-  , "short"
-  , "tiny"
-  , "colossal"
-  , "long"
-  , "short"
-  , "long"
-  , "gigantic"
-  ]
-
-ages :: [Text]
-ages =
-  [ "old"
-  , "young"
-  , "ancient"
-  , "elderly"
-  , "senior"
-  , "junior"
-  ]
-
-shapes :: [Text]
-shapes =
-  [ "wide"
-  , "narrow"
-  , "round"
-  , "triangular"
-  , "square"
-  , "flat"
-  , "thin"
-  , "thick"
-  , "skinny"
-  , "aerodynamic"
-  , "oval"
-  , "angular"
-  ]
-
-colours :: [Text]
-colours =
-  [ "red"
-  , "green"
-  , "blue"
-  , "yellow"
-  , "orange"
-  , "teal"
-  , "white"
-  , "black"
-  , "purple"
-  , "pink"
-  , "magenta"
-  , "cyan"
-  , "brown"
-  , "maroon"
-  , "aquamarine"
-  , "fuchsia"
-  , "crimson"
-  , "scarlet"
-  , "turquoise"
-  ]
-
-materials :: [Text]
-materials =
-  [ "glass"
-  , "wooden"
-  , "metalic"
-  , "leather"
-  , "polyester"
-  , "silk"
-  , "velvet"
-  , "nylon"
-  , "stone"
-  , "diamond"
-  , "plastic"
-  , "tin"
-  , "carbon"
-  , "cardboard"
-  , "paper"
-  , "sand"
-  , "plaster"
-  , "silicon"
-  , "canvas"
-  , "wool"
-  , "cotton"
-  , "marble"
-  ]
-
-nouns :: [Text]
-nouns =
-  [ "dragon"
-  , "unicorn"
-  , "mermaid"
-  , "fairy"
-  , "werewolf"
-  , "sphinx"
-  , "yeti"
-  , "griffin"
-  , "wolf"
-  , "crow"
-  , "centaur"
-  , "imp"
-  , "ghoul"
-  , "pixie"
-  , "gnome"
-  , "wizard"
-  , "witch"
-  , "mage"
-  , "troll"
-  , "cat"
-  , "dog"
-  , "snake"
-  , "lion"
-  , "monkey"
-  , "tiger"
-  , "fish"
-  , "crab"
-  , "shark"
-  , "salmon"
-  , "tuna"
-  , "horse"
-  , "turtle"
-  , "dolphin"
-  , "deer"
-  , "leopard"
-  , "bear"
-  , "frog"
-  , "llama"
-  , "penguin"
-  , "pig"
-  , "eagle"
-  , "bat"
-  , "vampire"
-  , "dinosaur"
-  , "whale"
-  , "king"
-  , "queen"
-  , "jester"
-  , "butterfly"
-  , "tulip"
-  , "polar-bear"
-  , "cactus"
-  , "hero"
-  , "knight"
-  , "hippogriff"
-  , "elf"
-  , "beast"
-  , "sprite"
-  , "alien"
-  , "ghost"
-  , "martian"
-  , "princess"
-  , "prince"
-  , "chef"
-  , "barista"
-  , "ufo"
-  ]
+    appName :: App.Name <- arbitrary
+    return . Subdomain $ textDisplay appName
