+++ fission-core/library/Fission/User/Username.hs
@@ -3,9 +3,7 @@
 module Fission.User.Username
   ( module Fission.User.Username.Types
   , module Fission.User.Username.Error
-  , module Fission.User.Username.Validation
   ) where
 
-import Fission.User.Username.Types
-import Fission.User.Username.Error
-import Fission.User.Username.Validation
+import           Fission.User.Username.Error
+import           Fission.User.Username.Types
