+++ fission-core/package.yaml
@@ -1,5 +1,5 @@
 name: fission-core
-version: '3.1.0.0'
+version: '3.2.0.0'
 category: API
 author:
   - Brooklyn Zelenka
