+++ fission-web-api/library/Fission/Web/API/App/Create/Types.hs
@@ -1,6 +1,7 @@
 module Fission.Web.API.App.Create.Types (Create) where
 
-import           Fission.URL
+import qualified Fission.App.Name.Types     as App
+import           Fission.URL.Types
 
 import           Fission.Web.API.Prelude
 
@@ -10,7 +11,7 @@ type Create
   =  Summary "Create app"
   :> Description "Creates a new app, assigns an initial subdomain, and sets an asset placeholder"
   --
-  :> QueryParam "subdomain" Subdomain
+  :> QueryParam "subdomain" App.Name
   --
   :> Auth.HigherOrder
   :> PostAccepted '[JSON] URL
