+++ fission-web-server/package.yaml
@@ -1,5 +1,5 @@
 name: fission-web-server
-version: '2.10.1.0'
+version: '2.11.0.0'
 category: API
 author:
   - Brooklyn Zelenka
